Added Stamina regen, Added Stamina Drain from weapons and movement

+++ The_Legend_Of_Oster/Assets/Scripts/Player/UI/HealthBar.cs
@@ -12,13 +12,13 @@ public class HealthBar : MonoBehaviour
         slider = GetComponent<Slider>();
     }
 
-    public void SetMaxHealth(int maxHealth)
+    public void SetMaxHealth(float maxHealth)
     {
         slider.maxValue = maxHealth;
         slider.value = maxHealth;
     }
 
-    public void SetCurrentHealth(int currentHealth)
+    public void SetCurrentHealth(float currentHealth)
     {
         slider.value = currentHealth;
     }
